Add authorization to routes

+++ back-end/src/routes/UserRoutes.kt
@@ -5,17 +5,23 @@ package com.gabriel.lunala.project.backend.routes
 import com.gabriel.lunala.project.backend.entities.map
 import com.gabriel.lunala.project.backend.services.UserService
 import io.ktor.application.*
+import io.ktor.config.*
 import io.ktor.http.*
 import io.ktor.locations.*
 import io.ktor.request.*
 import io.ktor.response.*
 import io.ktor.routing.*
+import io.ktor.util.*
+import io.ktor.util.pipeline.*
 
 const val PAGES_HEADER = "X-Total-Pages"
 
-@OptIn(KtorExperimentalLocationsAPI::class)
-fun Route.userRoutes(userService: UserService) {
+@OptIn(KtorExperimentalAPI::class, KtorExperimentalLocationsAPI::class)
+fun Route.userRoutes(appConfig: ApplicationConfig, userService: UserService) {
     put<Users.Id> { (id) ->
+        if (authorizate(appConfig).not())
+            return@put call.respond(HttpStatusCode.Unauthorized)
+
         when(val user = userService.updateById(id, call.receive())) {
             null -> call.respond(HttpStatusCode.NotFound)
             else -> call.respond(user.toResponseDto())
@@ -23,10 +29,16 @@ fun Route.userRoutes(userService: UserService) {
     }
 
     post<Users> {
+        if (authorizate(appConfig).not())
+            return@post call.respond(HttpStatusCode.Unauthorized)
+
         call.respond(HttpStatusCode.Created, userService.storeWithId(call.receive()).toResponseDto())
     }
 
     get<Users.Id> { (id) ->
+        if (authorizate(appConfig).not())
+            return@get call.respond(HttpStatusCode.Unauthorized)
+
         when (val user = userService.findById(id)) {
             null -> call.respond(HttpStatusCode.NotFound)
             else -> call.respond(user.toResponseDto())
@@ -34,6 +46,9 @@ fun Route.userRoutes(userService: UserService) {
     }
 
     get<Users.Paginated> { (page) ->
+        if (authorizate(appConfig).not())
+            return@get call.respond(HttpStatusCode.Unauthorized)
+
         val (pages, items) = userService.findPaginated(page).map { user ->
             user.toResponseDto()
         }
@@ -43,10 +58,18 @@ fun Route.userRoutes(userService: UserService) {
     }
 
     delete<Users.Id> { (id) ->
+        if (authorizate(appConfig).not())
+            return@delete call.respond(HttpStatusCode.Unauthorized)
+
         userService.deleteById(id)
         call.respond(HttpStatusCode.NoContent)
     }
+
 }
+
+@OptIn(KtorExperimentalAPI::class)
+fun PipelineContext<Unit, ApplicationCall>.authorizate(config: ApplicationConfig): Boolean =
+    config.property("keys").getList().contains(call.request.authorization())
 
 @Location("/users")
 class Users {
